requirements.txt
numpy
scikit-learn
flask
phe==1.5.0
requests
cryptography
tenseal
pyyaml
tensorflow
fastapi
boto3
jwt
uvicorn
torch
pandas
matplotlib
tensorflow_privacy